+++ akka-stream/src/main/scala/akka/stream/javadsl/FlattenStrategy.scala
@@ -3,7 +3,12 @@
  */
 package akka.stream.javadsl
 
-import akka.stream.javadsl
+import akka.stream.scaladsl
+
+/**
+ * Strategy that defines how a stream of streams should be flattened into a stream of simple elements.
+ */
+abstract class FlattenStrategy[-S, T] extends scaladsl.FlattenStrategy[S, T]
 
 object FlattenStrategy {
 
@@ -12,8 +17,11 @@ object FlattenStrategy {
    * emitting its elements directly to the output until it completes and then taking the next stream. This has the
    * consequence that if one of the input stream is infinite, no other streams after that will be consumed from.
    */
-  def concat[T]: akka.stream.FlattenStrategy[javadsl.Source[T, Unit], T] =
-    akka.stream.FlattenStrategy.Concat[T]().asInstanceOf[akka.stream.FlattenStrategy[javadsl.Source[T, _], T]]
-  // TODO so in theory this should be safe, but let's rethink the design later
+  def concat[T]: FlattenStrategy[Source[T, Unit], T] = Concat[T]()
+
+  /**
+   * INTERNAL API
+   */
+  private[akka] final case class Concat[T]() extends FlattenStrategy[Source[T, _], T]
 
 }
