+++ akka-stream/src/main/scala/akka/stream/scaladsl/Graph.scala
@@ -297,6 +297,9 @@ object FlowGraph extends GraphApply {
 
     private[stream] def module: Module = moduleInProgress
 
+    /** Converts this Scala DSL element to it's Java DSL counterpart. */
+    def asJava: javadsl.FlowGraph.Builder[M] = new javadsl.FlowGraph.Builder()(this)
+
   }
 
   object Implicits {
